Piggy bank simulator.
Now, you need to have a piggy bank with predefined capacity to hold your monies. The piggy bank holds the monies but gives them no other organization. And it can certainly contains duplicates. The piggy bank is an ADT that has only the following operations: 
•	check the capacity of the piggy bank,
•	you can add money to the piggy bank one at a time,
o	adding fails if there is no more room in the piggy bank (PiggyBankFullException will be thrown that must be handled by the client)
•	remove one (randomly selected),
o	remove fails if the piggy bank is empty (PiggyBankEmptyException will be thrown that must be handled by the client)
•	see whether the piggy bank is empty, 
•	see whether the piggy bank is full,
•	see how many monies can fit in the piggy bank,
•	see how many monies are in the piggy bank, 
•	see the content of the piggy bank,
•	shake the piggy bank to rearrange the monies,
•	empty the piggy bank counting how many monies landed HEADS